style: more cleanup
